Install powerline without --user if pyenv is found

+++ scripts/install-scripts/powerline.sh
@@ -5,7 +5,11 @@ expr "${0}" : "/.*" > /dev/null || cwd=`(cd "${cwd}" && pwd)`
 . $cwd/../lib.sh
 
 cyanecho ">>>> [installing powerline]"
-pip install --user git+git://github.com/Lokaltog/powerline
+if [ -e $HOME/.pyenv ]; then
+    pip install --user git+git://github.com/Lokaltog/powerline
+else
+    pip install git+git://github.com/Lokaltog/powerline
+fi
 
 cyanecho ">>>> [installing powerline font]"
 if [ `uname` == "Linux" ]; then
